+++ src/Data/Functor/Invariant.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE CPP #-}
+{-# OPTIONS_GHC -fno-warn-deprecations #-}
 
-#define GHC_GENERICS_OK __GLASGOW_HASKELL__ >= 706
+#define GHC_GENERICS_OK __GLASGOW_HASKELL__ >= 702
 
 {-|
 Module:      Data.Functor.Invariant
@@ -35,42 +36,98 @@ module Data.Functor.Invariant
   , WrappedProfunctor(..)
   ) where
 
-#if GHC_GENERICS_OK
-import GHC.Generics
-
-#endif
-import Text.ParserCombinators.ReadP (ReadP)
-import Text.ParserCombinators.ReadPrec (ReadPrec)
-
+-- base
 import qualified Control.Category as Cat
-import Control.Arrow
-import Control.Applicative as App
-import Control.Monad (MonadPlus(..))
-import Control.Monad.ST (ST)
+import           Control.Arrow
+import           Control.Applicative as App
+import           Control.Exception (Handler(..))
+import           Control.Monad (MonadPlus(..))
+import qualified Control.Monad.ST as Strict (ST)
+import qualified Control.Monad.ST.Lazy as Lazy (ST)
+import           Data.Functor.Identity (Identity)
+import           Data.Ix (Ix)
+import qualified Data.Monoid as Monoid (First(..), Last(..))
+#if MIN_VERSION_base(4,8,0)
+import           Data.Monoid (Alt(..))
+#endif
+import           Data.Monoid (Dual(..), Endo(..))
+import           Data.Proxy (Proxy(..))
+#if GHC_GENERICS_OK
+import           GHC.Generics
+#endif
+import           System.Console.GetOpt as GetOpt
+import           Text.ParserCombinators.ReadP (ReadP)
+import           Text.ParserCombinators.ReadPrec (ReadPrec)
 
-import Data.Bifunctor hiding (first)
-import Data.Bifunctor.Biff
-import Data.Bifunctor.Clown
-import Data.Bifunctor.Flip
-import Data.Bifunctor.Join
-import Data.Bifunctor.Joker
-import Data.Bifunctor.Product
-import Data.Bifunctor.Tannen
-import Data.Bifunctor.Wrapped
+-- array
+import           Data.Array (Array)
 
-import Data.Functor.Contravariant
-import Data.Functor.Contravariant.Compose
-import Data.Functor.Contravariant.Divisible
+-- bifunctors
+import           Data.Bifunctor hiding (first)
+import           Data.Bifunctor.Biff
+import           Data.Bifunctor.Clown
+import           Data.Bifunctor.Flip
+import           Data.Bifunctor.Join
+import           Data.Bifunctor.Joker
+import qualified Data.Bifunctor.Product as Bifunctors
+import           Data.Bifunctor.Tannen
+import           Data.Bifunctor.Wrapped
 
-import Data.Monoid (Dual(Dual), Endo(Endo))
+-- containers
+import           Data.IntMap (IntMap)
+import           Data.Map (Map)
+import           Data.Sequence (Seq, ViewL, ViewR)
+import           Data.Tree (Tree)
 
-import Data.Profunctor as Pro
-import Data.Profunctor.Cayley
-import Data.Profunctor.Closed
-import Data.Profunctor.Codensity
-import Data.Profunctor.Composition
-import Data.Profunctor.Ran
-import Data.Profunctor.Tambara
+-- contravariant
+import           Data.Functor.Contravariant
+import           Data.Functor.Contravariant.Compose as Contravariant
+import           Data.Functor.Contravariant.Divisible
+
+-- profunctors
+import           Data.Profunctor as Pro
+import           Data.Profunctor.Cayley
+import           Data.Profunctor.Closed
+import           Data.Profunctor.Codensity
+import           Data.Profunctor.Composition
+import           Data.Profunctor.Ran
+import           Data.Profunctor.Tambara
+
+-- semigroups
+import           Data.List.NonEmpty (NonEmpty(..))
+import qualified Data.Semigroup as Semigroup (First(..), Last(..), Option(..))
+import           Data.Semigroup (Min(..), Max(..), Arg(..))
+
+-- stm
+import           Control.Concurrent.STM (STM)
+
+-- tagged
+import           Data.Tagged (Tagged(..))
+
+-- transformers
+import           Control.Applicative.Backwards (Backwards(..))
+import           Control.Applicative.Lift (Lift(..))
+import           Control.Monad.Trans.Cont (ContT)
+import           Control.Monad.Trans.Error (ErrorT(..))
+import           Control.Monad.Trans.Except (ExceptT(..), runExceptT)
+import           Control.Monad.Trans.Identity (IdentityT, mapIdentityT)
+import           Control.Monad.Trans.List (ListT, mapListT)
+import           Control.Monad.Trans.Maybe (MaybeT, mapMaybeT)
+import qualified Control.Monad.Trans.RWS.Lazy as Lazy (RWST(..))
+import qualified Control.Monad.Trans.RWS.Strict as Strict (RWST(..))
+import           Control.Monad.Trans.Reader (ReaderT, mapReaderT)
+import qualified Control.Monad.Trans.State.Lazy as Lazy (StateT(..))
+import qualified Control.Monad.Trans.State.Strict as Strict (StateT(..))
+import qualified Control.Monad.Trans.Writer.Lazy as Lazy (WriterT, mapWriterT)
+import qualified Control.Monad.Trans.Writer.Strict as Strict (WriterT, mapWriterT)
+import qualified Data.Functor.Compose as Transformers (Compose(..))
+import           Data.Functor.Constant (Constant(..))
+import           Data.Functor.Product as Transformers (Product(..))
+import           Data.Functor.Reverse (Reverse(..))
+import           Data.Functor.Sum as Transformers (Sum(..))
+
+-- unordered-containers
+import           Data.HashMap.Lazy (HashMap)
 
 -------------------------------------------------------------------------------
 -- The Invariant class
@@ -101,7 +158,8 @@ invmapContravariant = const contramap
 instance Invariant Maybe where invmap = invmapFunctor
 instance Invariant [] where invmap = invmapFunctor
 instance Invariant IO where invmap = invmapFunctor
-instance Invariant (ST s) where invmap = invmapFunctor
+instance Invariant (Strict.ST s) where invmap = invmapFunctor
+instance Invariant (Lazy.ST s) where invmap = invmapFunctor
 instance Invariant ReadP where invmap = invmapFunctor
 instance Invariant ReadPrec where invmap = invmapFunctor
 instance Invariant ((->) a) where invmap = invmapFunctor
@@ -123,11 +181,66 @@ instance Monad m => Invariant (WrappedMonad m) where invmap = invmapFunctor
 instance Arrow arr => Invariant (App.WrappedArrow arr a) where
   invmap f _ (App.WrapArrow x) = App.WrapArrow $ ((arr f) Cat.. x)
 
+-- | from @Control.Arrow@
+instance
+#if MIN_VERSION_base(4,4,0)
+  Arrow a
+#else
+  ArrowApply a
+#endif
+  => Invariant (ArrowMonad a) where
+  invmap f _ (ArrowMonad m) = ArrowMonad $ m >>> arr f
+
+-- | from @Control.Exception@
+instance Invariant Handler where
+  invmap f _ (Handler h) = Handler (fmap f . h)
+
+-- | from @Data.Functor.Identity@
+instance Invariant Identity where
+  invmap = invmapFunctor
+
 -- | from @Data.Monoid@
 instance Invariant Dual where invmap f _ (Dual x) = Dual (f x)
 -- | from @Data.Monoid@
 instance Invariant Endo where
   invmap f g (Endo x) = Endo (f . x . g)
+-- | from @Data.Monoid@
+instance Invariant Monoid.First where
+  invmap f g (Monoid.First x) = Monoid.First (invmap f g x)
+-- | from @Data.Monoid@
+instance Invariant Monoid.Last where
+  invmap f g (Monoid.Last x) = Monoid.Last (invmap f g x)
+#if MIN_VERSION_base(4,8,0)
+-- | from @Data.Monoid@
+instance Invariant f => Invariant (Alt f) where
+  invmap f g (Alt x) = Alt (invmap f g x)
+#endif
+
+-- | from @Data.Proxy@
+instance Invariant Proxy where
+  invmap = invmapFunctor
+
+-- | from @System.Console.GetOpt@
+instance Invariant ArgDescr where
+  invmap f _ (NoArg a)    = NoArg (f a)
+  invmap f _ (ReqArg g s) = ReqArg (f . g) s
+  invmap f _ (OptArg g s) = OptArg (f . g) s
+-- | from @System.Console.GetOpt@
+instance Invariant ArgOrder where
+  invmap _ _ RequireOrder      = RequireOrder
+  invmap _ _ Permute           = Permute
+  invmap f _ (ReturnInOrder g) = ReturnInOrder (f . g)
+-- | from @System.Console.GetOpt@
+instance Invariant OptDescr where
+  invmap f g (GetOpt.Option a b argDescr c) = GetOpt.Option a b (invmap f g argDescr) c
+
+-- | from the @array@ package
+instance
+#if __GLASGOW_HASKELL__ < 711
+  Ix i =>
+#endif
+    Invariant (Array i) where
+  invmap = invmapFunctor
 
 -- | from the @bifunctors@ package
 instance (Invariant2 p, Invariant g) => Invariant (Biff p f g a) where
@@ -151,6 +264,25 @@ instance (Invariant f, Invariant2 p) => Invariant (Tannen f p a) where
 instance Bifunctor p => Invariant (WrappedBifunctor p a) where
   invmap = invmap2 id id
 
+-- | from the @containers@ package
+instance Invariant IntMap where
+  invmap = invmapFunctor
+-- | from the @containers@ package
+instance Invariant (Map k) where
+  invmap = invmapFunctor
+-- | from the @containers@ package
+instance Invariant Seq where
+  invmap = invmapFunctor
+-- | from the @containers@ package
+instance Invariant ViewL where
+  invmap = invmapFunctor
+-- | from the @containers@ package
+instance Invariant ViewR where
+  invmap = invmapFunctor
+-- | from the @containers@ package
+instance Invariant Tree where
+  invmap = invmapFunctor
+
 -- | from the @contravariant@ package
 instance Invariant Predicate where invmap = invmapContravariant
 -- | from the @contravariant@ package
@@ -160,8 +292,9 @@ instance Invariant Equivalence where invmap = invmapContravariant
 -- | from the @contravariant@ package
 instance Invariant (Op a) where invmap = invmapContravariant
 -- | from the @contravariant@ package
-instance (Invariant f, Invariant g) => Invariant (Compose f g) where
-  invmap f g (Compose x) = Compose $ invmap (invmap f g) (invmap g f) x
+instance (Invariant f, Invariant g) => Invariant (Contravariant.Compose f g) where
+  invmap f g (Contravariant.Compose x) =
+    Contravariant.Compose $ invmap (invmap f g) (invmap g f) x
 -- | from the @contravariant@ package
 instance (Invariant f, Invariant g) => Invariant (ComposeCF f g) where
   invmap f g (ComposeCF x) = ComposeCF $ invmap (invmap f g) (invmap g f) x
@@ -202,6 +335,114 @@ instance Invariant2 p => Invariant (Tambara p a) where
 -- | from the @profunctors@ package
 instance Invariant2 p => Invariant (Cotambara p a) where
   invmap = invmap2 id id
+
+-- | from the @semigroups@ package
+instance Invariant NonEmpty where
+  invmap = invmapFunctor
+-- | from the @semigroups@ package
+instance Invariant Min where
+  invmap = invmapFunctor
+-- | from the @semigroups@ package
+instance Invariant Max where
+  invmap = invmapFunctor
+-- | from the @semigroups@ package
+instance Invariant Semigroup.First where
+  invmap = invmapFunctor
+-- | from the @semigroups@ package
+instance Invariant Semigroup.Last where
+  invmap = invmapFunctor
+-- | from the @semigroups@ package
+instance Invariant Semigroup.Option where
+  invmap = invmapFunctor
+-- | from the @semigroups@ package
+instance Invariant (Arg a) where
+  invmap = invmapFunctor
+
+-- | from the @stm@ package
+instance Invariant STM where
+  invmap = invmapFunctor
+
+-- | from the @tagged@ package
+instance Invariant (Tagged s) where
+  invmap = invmapFunctor
+
+-- | from the @transformers@ package
+instance Invariant f => Invariant (Backwards f) where
+  invmap f g (Backwards a) = Backwards (invmap f g a)
+-- | from the @transformers@ package
+instance Invariant f => Invariant (Lift f) where
+  invmap f _ (Pure x)  = Pure (f x)
+  invmap f g (Other y) = Other (invmap f g y)
+-- | from the @transformers@ package
+instance Invariant (ContT r m) where
+  invmap = invmapFunctor
+-- -- | from the @transformers@ package
+instance Invariant m => Invariant (ErrorT e m) where
+  invmap f g = ErrorT . invmap (invmap f g) (invmap g f) . runErrorT
+-- | from the @transformers@ package
+instance Invariant m => Invariant (ExceptT e m) where
+  invmap f g = ExceptT . invmap (invmap f g) (invmap g f) . runExceptT
+-- | from the @transformers@ package
+instance Invariant m => Invariant (IdentityT m) where
+  invmap f g = mapIdentityT (invmap f g)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (ListT m) where
+  invmap f g = mapListT $ invmap (invmap f g) (invmap g f)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (MaybeT m) where
+  invmap f g = mapMaybeT $ invmap (invmap f g) (invmap g f)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (Lazy.RWST r w s m) where
+  invmap f g m = Lazy.RWST $ \r s ->
+    invmap (mapFstTriple f) (mapFstTriple g) $ Lazy.runRWST m r s
+      where mapFstTriple h ~(a, s, w) = (h a, s, w)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (Strict.RWST r w s m) where
+  invmap f g m = Strict.RWST $ \r s ->
+    invmap (mapFstTriple f) (mapFstTriple g) $ Strict.runRWST m r s
+      where mapFstTriple h (a, s, w) = (h a, s, w)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (ReaderT r m) where
+  invmap f g = mapReaderT (invmap f g)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (Lazy.StateT s m) where
+  invmap f g m = Lazy.StateT $ \s ->
+    invmap (mapFstPair f) (mapFstPair g) $ Lazy.runStateT m s
+      where mapFstPair h ~(a, s) = (h a, s)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (Strict.StateT s m) where
+  invmap f g m = Strict.StateT $ \s ->
+    invmap (mapFstPair f) (mapFstPair g) $ Strict.runStateT m s
+      where mapFstPair h (a, s) = (h a, s)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (Lazy.WriterT w m) where
+  invmap f g = Lazy.mapWriterT $ invmap (mapFstPair f) (mapFstPair g)
+    where mapFstPair h ~(a, w) = (h a, w)
+-- | from the @transformers@ package
+instance Invariant m => Invariant (Strict.WriterT w m) where
+  invmap f g = Strict.mapWriterT $ invmap (mapFstPair f) (mapFstPair g)
+    where mapFstPair h (a, w) = (h a, w)
+-- | from the @transformers@ package
+instance (Invariant f, Invariant g) => Invariant (Transformers.Compose f g) where
+  invmap f g (Transformers.Compose x) =
+    Transformers.Compose (invmap (invmap f g) (invmap g f) x)
+-- | from the @transformers@ package
+instance Invariant (Constant a) where
+  invmap = invmapFunctor
+-- | from the @transformers@ package
+instance (Invariant f, Invariant g) => Invariant (Transformers.Product f g) where
+  invmap f g (Transformers.Pair x y) = Transformers.Pair (invmap f g x) (invmap f g y)
+-- | from the @transformers@ package
+instance Invariant f => Invariant (Reverse f) where
+  invmap f g (Reverse a) = Reverse (invmap f g a)
+-- | from the @transformers@ package
+instance (Invariant f, Invariant g) => Invariant (Transformers.Sum f g) where
+  invmap f g (InL x) = InL (invmap f g x)
+  invmap f g (InR y) = InR (invmap f g y)
+
+-- | from the @unordered-containers@ package
+instance Invariant (HashMap k) where
+  invmap = invmapFunctor
 
 -------------------------------------------------------------------------------
 -- WrappedFunctor
@@ -248,12 +489,14 @@ instance Contravariant f => Contravariant (WrappedContravariant f) where
   contramap f = WrapContravariant . contramap f . unwrapContravariant
 
 instance Divisible f => Divisible (WrappedContravariant f) where
-  divide f (WrapContravariant l) (WrapContravariant r) = WrapContravariant $ divide f l r
+  divide f (WrapContravariant l) (WrapContravariant r) =
+    WrapContravariant $ divide f l r
   conquer = WrapContravariant conquer
 
 instance Decidable f => Decidable (WrappedContravariant f) where
   lose = WrapContravariant . lose
-  choose f (WrapContravariant l) (WrapContravariant r) = WrapContravariant $ choose f l r
+  choose f (WrapContravariant l) (WrapContravariant r) =
+    WrapContravariant $ choose f l r
 
 -------------------------------------------------------------------------------
 -- The Invariant2 class
@@ -305,7 +548,8 @@ instance Arrow arr => Invariant2 (App.WrappedArrow arr) where
 
 -- | from the @bifunctors@ package
 instance (Invariant2 p, Invariant f, Invariant g) => Invariant2 (Biff p f g) where
-  invmap2 f f' g g' = Biff . invmap2 (invmap f f') (invmap f' f) (invmap g g') (invmap g' g) . runBiff
+  invmap2 f f' g g' =
+    Biff . invmap2 (invmap f f') (invmap f' f) (invmap g g') (invmap g' g) . runBiff
 -- | from the @bifunctors@ package
 instance Invariant f => Invariant2 (Clown f) where
   invmap2 f f' _ _ = Clown . invmap f f' . runClown
@@ -316,11 +560,13 @@ instance Invariant2 p => Invariant2 (Flip p) where
 instance Invariant g => Invariant2 (Joker g) where
   invmap2 _ _ g g' = Joker . invmap g g' . runJoker
 -- | from the @bifunctors@ package
-instance (Invariant2 f, Invariant2 g) => Invariant2 (Product f g) where
-  invmap2 f f' g g' (Pair x y) = Pair (invmap2 f f' g g' x) (invmap2 f f' g g' y)
+instance (Invariant2 f, Invariant2 g) => Invariant2 (Bifunctors.Product f g) where
+  invmap2 f f' g g' (Bifunctors.Pair x y) =
+    Bifunctors.Pair (invmap2 f f' g g' x) (invmap2 f f' g g' y)
 -- | from the @bifunctors@ package
 instance (Invariant f, Invariant2 p) => Invariant2 (Tannen f p) where
-  invmap2 f f' g g' = Tannen . invmap (invmap2 f f' g g') (invmap2 f' f g' g) . runTannen
+  invmap2 f f' g g' =
+    Tannen . invmap (invmap2 f f' g g') (invmap2 f' f g' g) . runTannen
 -- | from the @bifunctors@ package
 instance Bifunctor p => Invariant2 (WrappedBifunctor p) where
   invmap2 f f' g g' = WrapBifunctor . invmap2Bifunctor f f' g g' . unwrapBifunctor
@@ -343,7 +589,8 @@ instance Invariant2 (Forget r) where
   invmap2 = invmap2Profunctor
 -- | from the @profunctors@ package
 instance (Invariant f, Invariant2 p) => Invariant2 (Cayley f p) where
-  invmap2 f f' g g' = Cayley . invmap (invmap2 f f' g g') (invmap2 f' f g' g) . runCayley
+  invmap2 f f' g g' =
+    Cayley . invmap (invmap2 f f' g g') (invmap2 f' f g' g) . runCayley
 -- | from the @profunctors@ package
 instance Invariant2 p => Invariant2 (Closure p) where
   invmap2 f f' g g' (Closure p) = Closure $ invmap2 (f .) (f' .) (g .) (g' .) p
@@ -352,10 +599,12 @@ instance Invariant2 (Environment p) where
   invmap2 _ f' g _ (Environment l m r) = Environment (g . l) m (r . f')
 -- | from the @profunctors@ package
 instance Invariant2 p => Invariant2 (Codensity p) where
-  invmap2 ac ca bd db (Codensity f) = Codensity (invmap2 id id bd db . f . invmap2 id id ca ac)
+  invmap2 ac ca bd db (Codensity f) =
+    Codensity (invmap2 id id bd db . f . invmap2 id id ca ac)
 -- | from the @profunctors@ package
 instance (Invariant2 p, Invariant2 q) => Invariant2 (Procompose p q) where
-  invmap2 l l' r r' (Procompose f g) = Procompose (invmap2 id id r r' f) (invmap2 l l' id id g)
+  invmap2 l l' r r' (Procompose f g) =
+    Procompose (invmap2 id id r r' f) (invmap2 l l' id id g)
 -- | from the @profunctors@ package
 instance (Invariant2 p, Invariant2 q) => Invariant2 (Rift p q) where
   invmap2 ac ca bd db (Rift f) = Rift (invmap2 ac ca id id . f . invmap2 db bd id id)
@@ -364,16 +613,30 @@ instance (Invariant2 p, Invariant2 q) => Invariant2 (Ran p q) where
   invmap2 ac ca bd db (Ran f) = Ran (invmap2 id id bd db . f . invmap2 id id ca ac)
 -- | from the @profunctors@ package
 instance Invariant2 p => Invariant2 (Tambara p) where
-  invmap2 f f' g g' (Tambara p) = Tambara $ invmap2 (first f) (first f') (first g) (first g') p
+  invmap2 f f' g g' (Tambara p) =
+    Tambara $ invmap2 (first f) (first f') (first g) (first g') p
 -- | from the @profunctors@ package
 instance Invariant2 (Pastro p) where
   invmap2 _ f' g _ (Pastro l m r) = Pastro (g . l) m (r . f')
 -- | from the @profunctors@ package
 instance Invariant2 p => Invariant2 (Cotambara p) where
-  invmap2 f f' g g' (Cotambara p) = Cotambara $ invmap2 (left f) (left f') (left g) (left g') p
+  invmap2 f f' g g' (Cotambara p) =
+    Cotambara $ invmap2 (left f) (left f') (left g) (left g') p
 -- | from the @profunctors@ package
 instance Invariant2 (Copastro p) where
   invmap2 _ f' g _ (Copastro l m r) = Copastro (g . l) m (r . f')
+
+-- | from the @semigroups@ package
+instance Invariant2 Arg where
+  invmap2 = invmap2Bifunctor
+
+-- | from the @tagged@ package
+instance Invariant2 Tagged where
+  invmap2 = invmap2Bifunctor
+
+-- | from the @transformers@ package
+instance Invariant2 Constant where
+  invmap2 f _ _ _ (Constant x) = Constant (f x)
 
 -------------------------------------------------------------------------------
 -- WrappedProfunctor
@@ -410,11 +673,11 @@ instance Cochoice p => Cochoice (WrappedProfunctor p) where
 
 instance Closed p => Closed (WrappedProfunctor p) where
   closed = WrapProfunctor . closed . unwrapProfunctor
+
 #if GHC_GENERICS_OK
-
-
-
-
+-------------------------------------------------------------------------------
+-- GHC Generics
+-------------------------------------------------------------------------------
 
 -- | from @GHC.Generics@
 instance Invariant V1 where
@@ -443,9 +706,6 @@ instance Invariant f => Invariant (Rec1 f) where invmap f g (Rec1 fp) = Rec1 $ i
 -- likely requires writing your 'Generic1' instance by hand
 instance (Invariant f, Invariant g) => Invariant ((:.:) f g) where
   invmap f g (Comp1 fgp) = Comp1 $ invmap (invmap f g) (invmap g f) fgp
-
-
-
 
 
 {- $ghcgenerics