+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 * Added `WrappedFunctor`, `WrappedContravariant`, `WrappedBifunctor`, and
   `WrappedProfunctor` data types to allow use of `invmap` and `invmap2` for
   data types that aren't `Invariant` or `Invariant2` instances.
-* Added `Invariant` and `Invariant2` instances for data types in the
-  `bifunctors` and `profunctors` libraries
+* Added `Invariant` instances for lazy `ST`, `ArrowMonad`, `Handler`,
+  `Identity`, `First`, `Last`, `Alt`, `Proxy`, `ArgDescr`, `ArgOrder`, and
+  `OptDescr`
+* Added `Invariant` and `Invariant2` instances for data types in the `array`,
+  `bifunctors`, `containers`, `profunctors`, `semigroups`, `stm`, `tagged`,
+  `transformers`, and `unordered-containers` libraries
 
 # 0.1.2
 * Add `Invariant` instances for `Dual` and `Endo`